Fixed physics differences between HTML5 and Windows
The game now spawns different objects based on whether the game is being run in a browser or not. The difference between the objects is that the Windows block object uses the "Uses Physics" checkbox and has positive values for the collision mask offset whereas the box is unchecked and negative values are used for the browser block object.